refactor: move util into source
