+++ deploy/deploy_gradio.py
@@ -12,12 +12,12 @@ from PIL import Image
 from pathlib import Path
 from tqdm import tqdm
 
-from src.craftdet.detection import Detector
+
 from vietocr.tool.predictor import Predictor
 from vietocr.tool.config import Cfg
+from src.craftdet.detection import Detector
 from src.preprocessor.model import DewarpTextlineMaskGuide
-
-from .utils import pdf2imgs, bbox2ibox, cv2crop, cv2drawbox
+from src.utils import pdf2imgs, bbox2ibox, cv2crop, cv2drawbox
 
 DEFAULT_SIZE_IMAGE = 224
 save_origin_path = Path(os.getcwd() + "/prediction/origin").expanduser().resolve()
