+++ deploy/test_rectification.py
@@ -16,11 +16,12 @@ from PIL import Image
 from pathlib import Path
 from tqdm import tqdm
 
-from src.craftdet.detection import Detector
+
 from vietocr.tool.predictor import Predictor
 from vietocr.tool.config import Cfg
+from src.craftdet.detection import Detector
 from src.preprocessor.model import DewarpTextlineMaskGuide
-from deploy.utils import pdf2imgs, bbox2ibox, cv2crop, cv2drawbox
+from src.utils import pdf2imgs, bbox2ibox, cv2crop, cv2drawbox
 
 
 def get_args():
